changed the logic of retrieving user id
