+++ src/main/java/tj/alimov/productservice/service/JwtService.java
@@ -7,6 +7,7 @@ import io.jsonwebtoken.security.Keys;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import tj.alimov.productservice.exception.user.TokenNotProvidedException;
 
 import java.nio.charset.StandardCharsets;
 import java.security.Key;
@@ -37,8 +38,22 @@ public class JwtService {
                 .parseClaimsJws(token)
                 .getBody();
     }
+
+    public Long getUserId(HttpServletRequest request){
+        String token = retrieveToken(request);
+        return extractUserId(token);
+    }
     public Long extractUserId(String token){
         Claims claims = extractClaims(token);
         return claims.get("userId", Long.class);
+    }
+
+    private String retrieveToken(HttpServletRequest request){
+        String header = request.getHeader("Authentication");
+        if(header == null || !header.startsWith("Bearer")){
+            throw new TokenNotProvidedException("Token is not provided");
+        }
+        String token = header.substring("Bearer ".length());
+        return token;
     }
 }
